requirements.txt
Django==5.0.6
django-extensions==3.2.3
django-filter==24.2
django-simple-history==3.5.0
djangorestframework==3.15.1
django-cors-headers==4.3.1
Markdown==3.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
redis==5.0.4
celery==5.4.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
